# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Benetnasch API
  version: 1.0.0
  description: Versioned HTTP contract for the Benetnasch blog and administration apps.
servers:
  - url: /api/v1
...
paths:
...
  /public/articles/{articleId}:
    get: { tags: [public], summary: Get article detail, operationId: getArticle, parameters: [{ $ref: '#/components/parameters/ArticleId' }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /public/articles/{articleId}/access:
    post: { tags: [public], summary: Verify article password, operationId: accessArticle, parameters: [{ $ref: '#/components/parameters/ArticleId' }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/articles:
    get: { tags: [admin], summary: List admin articles, operationId: listAdminArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save an article, operationId: saveAdminArticle, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/{articleId}:
    get: { tags: [admin], summary: Get admin article, operationId: getAdminArticle, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ArticleId' }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/images:
    post: { tags: [admin], summary: Upload article image, operationId: uploadArticleImage, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/import:
    post: { tags: [admin], summary: Import articles, operationId: importArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/export:
    post: { tags: [admin], summary: Export articles, operationId: exportArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/featured:
    put: { tags: [admin], summary: Update article flags, operationId: updateArticleFlags, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/trash:
    put: { tags: [admin], summary: Move articles to or from trash, operationId: updateArticleTrash, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/batch-delete:
    delete: { tags: [admin], summary: Permanently delete articles, operationId: deleteArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/categories:
    get: { tags: [admin], summary: List admin categories, operationId: listAdminCategories, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save category, operationId: saveCategory, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete categories, operationId: deleteCategories, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/categories/search:
    get: { tags: [admin], summary: Search admin categories, operationId: searchAdminCategories, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/albums:
    get: { tags: [admin], summary: List albums, operationId: listAdminAlbums, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save album, operationId: saveAlbum, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/albums/{albumId}:
    get: { tags: [admin], summary: Get album, operationId: getAdminAlbum, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/AlbumId' }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete album, operationId: deleteAlbum, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/AlbumId' }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/albums/options:
    get: { tags: [admin], summary: List album options, operationId: listAlbumOptions, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/albums/cover:
    post: { tags: [admin], summary: Upload album cover, operationId: uploadAlbumCover, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/roles:
    get: { tags: [admin], summary: List roles, operationId: listRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save role, operationId: saveRole, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete roles, operationId: deleteRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/users/roles:
    get: { tags: [admin], summary: List user role options, operationId: listUserRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    put: { tags: [admin], summary: Update user roles, operationId: updateUserRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/logs/jobs:
    get: { tags: [admin], summary: List job logs, operationId: listJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete job logs, operationId: deleteJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/jobs/clean:
    delete: { tags: [admin], summary: Clean job logs, operationId: cleanJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/jobs/groups:
    get: { tags: [admin], summary: List job log groups, operationId: listJobLogGroups, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/logs/exceptions:
    get: { tags: [admin], summary: List exception logs, operationId: listExceptionLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete exception logs, operationId: deleteExceptionLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
components:
  parameters:
    ArticleId: { name: articleId, in: path, required: true, schema: { type: integer, format: int64 } }
    AlbumId: { name: albumId, in: path, required: true, schema: { type: integer, format: int64 } }
...
  responses:
    Result:
      description: Versioned API response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: 操作成功 }
        data: { nullable: true }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT